build: refactored visaulizations
